backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
xlrd
pypdf
python-docx
sentence-transformers
google-generativeai
python-dotenv
numpy
chromadb
